+++ packages/lucid/src/tx-builder/internal/Mint.ts
@@ -1,7 +1,7 @@
 import { Effect, pipe } from "effect";
 import { toText } from "@lucid-evolution/core-utils";
 import { Assets, Redeemer } from "@lucid-evolution/core-types";
-import * as CML from "@dcspark/cardano-multiplatform-lib-nodejs";
+import * as CML from "@anastasia-labs/cardano-multiplatform-lib-nodejs";
 import { toPartial, toV1, toV2 } from "./TxUtils.js";
 import {
   ERROR_MESSAGE,